+++ src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java
@@ -1,6 +1,5 @@
 package com.blackbuild.groovy.configdsl.transform;
 
-import org.codehaus.groovy.transform.CanonicalASTTransformation;
 import org.codehaus.groovy.transform.GroovyASTTransformationClass;
 
 import java.lang.annotation.*;
@@ -8,8 +7,8 @@ import java.lang.annotation.*;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.SOURCE)
 @Inherited // This is currently not used, see https://issues.apache.org/jira/browse/GROOVY-6765
-@GroovyASTTransformationClass(classes={DSLConfigASTTransformation.class})
-public @interface DSLConfig {
+@GroovyASTTransformationClass(classes={DSLASTTransformation.class})
+public @interface DSL {
 
     /**
      * Which field of this class is used as the key when an instance of this class is used in a collection.