+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/model/DSLDemoSpec.groovy
@@ -7,14 +7,14 @@ class DSLDemoSpec extends AbstractDSLSpec {
         createClass('''
             package com.blackbuild.groovy.configdsl.transform.test
 
-            @DSLConfig
+            @DSL
             class Config {
                 String name
 
                 Map<String, Environment> envs
             }
 
-            @DSLConfig(key = "name")
+            @DSL(key = "name")
             class Environment {
                 String name
                 String value