+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/model/MetaSpec.groovy
@@ -13,7 +13,7 @@ class MetaSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
             }
@@ -34,7 +34,7 @@ class MetaSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL(key = "name")
             class Foo {
                 String name
                 String value