+++ CHANGES.md
@@ -1,3 +1,11 @@
+pending
+=======
+- Major renaming of annotations:
+  - DSLConfig -> DSL
+  - DSLField -> Field
+  
+
+
 0.11.0
 ======
 - Introduced template objects to configure default values.